Fix for Chrome for Android
Chrome for Android resizes the browser a lot when scrolling due to the
address bar collapsing. This causes a lot of arbitrary layout jumps and
slow image resize operations with this plugin.

So for device UA where height should not change, we only update if the
width changes as well (f.ex. orientation changes).